refactor: 구매한 로또 번호와 지난주 당첨 번호 비교 로직을 Lotteries로 이동
