+++ README.md
@@ -142,7 +142,7 @@ ___
 - [x] 수동으로 구매할 번호 입력
   - [x] 수동 로또 생성
 - [x] Printer %s, %f 사용
-- [ ] LotteryComparator를 Lotteries로 합칠 것이냐
+- [x] LotteryComparator 기능을 Lotteries로 합치기
 - [ ] PrizeMoney로 당첨 횟수 세기
 
 ### 논의거리
